pyproject.toml
[project]
name = "repl-nix-pdfcompilator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.7.1",
    "pillow>=11.0.0",
    "pypdf2>=3.0.1",
    "pyqt5>=5.15.11",
]
